UI: Fade in Region and Area Move Divider Lines
When resizing areas and regions there is a line shown along the edge to
help indicate that you have pressed the correct location. This PR just
quickly animates this line into place, rather than immediately. With
area resizing the line fades in and widens. With region resizing it
also moves away from the edge if minimized.