Fix #140642: EEVEE: Use overriden world to sync volumes
Because `World::sync_volume()`` takes place before the rest of the world
is sync, if the overriden world has different volume setup than the
original, the rendered volume result may not be expected. This fix uses
overriden world for `sync_volume()` to mitigate this issue.

Also fixes the Material/World override panel visibility issue when in
EEVEE render.