Fix long running loading setting status to loading after timeout
If the loading would run over the timeout threshold and still call
`asset_library_status_is_loading()`, the library status would be set
back to "loading". Seperate the function to reset the timeout from the
function that marks the begin of loading. The latter function is the
only one changing the state to loading from now on.