Fix #140700: Blender 4.5 Crashes When Performing Loop Cut
Correct an oversight: without proportional editing, all the transform
data is "selected" by definition. This means that the assumption "the
selected data is sorted first in the array" is trivially true, without
calling any sorting function. So instead of using the sorted index map
in these cases, just fall back to regular iteration.

To make the code handle this nicely, I made two "foreach" functions. The
first one transparently uses the sorted index map when available, and
performs regular iteration otherwise. The second function only visits
the selected items.

This makes the usage of these functions clearer, and the fact that
selected items are expected to be sorted first (either trivially or
explicitly) can be documented in a central place.