Fix #140574: Manifold boolean leaves redundant 2-edged vertices.
The triangulation edges can be cut by other other objects and then
if we dissolve those edges, this cane leave an "on edge" vertex
that is in line with its two ends, and is not needed.
The fix is to detect such vertices and delete them from the output,
closing up the affected faces.
For some reason I don't understand yet, this didn't happen with the
version 3.0.1 Manifold library, but does with the 3.1.0 one.