UI: Reduce text jumping for header status
Many operators in the 3D viewport print values with a fixed number of
decimal places using format specifiers like `%.`, where jumping is
almost nonexistent.

Problems only arise with operators that use `BKE_unit_value_as_string`
or its derivatives, since the function requests precision in terms of
significant figures, which means the length of the fractional part
frequently jumps around.

To mitigate this issue, we can pass a negative precision which disables
the stripping of trailing zeros (added in f69c565a33).

This patch intends to build on #139507, but is useful even without it.