Match starting number of decimals to the limit, reducing initial jump
This is only enabled if the precision passed in is negative, meaning it does not affect any existing users of the function unless they opt in.

Resulting transition is very nice + smooth since the length of the string starts long and gets shorter the less precise you go, as opposed to before where it would jump from a short -> long -> short string.