UI Experiment: Fade In Active Editor Border Highlight
Our themes have two colors that we use as outlines for the editors, so
that we can differentiate the active area. This PR just fades in the
"Active Editor Outline" color rather than show it instantly. It makes
the change a little less jarring.  If we like this it could be improved
by also transitioning out the active color from the old area, but this
is not done here yet.