Fix #140801: Particle Weight Overlay Broken
Was not showing weights on the lines (but the points instead).
Points were supposed to show selection though, not weights.

This is now corrected.

NOTE: selection itself is broken as well (`Select All` operator shows
the selection is drawing fine, but it seems depths are not drawing
correctly - which will mess up `key_test_depth`...), but this is
probably for another commit.