Fix #63595: Generated texture coordinates missing when switching from solid to rendered mode
The issue is caused by `ED_view3d_datamask` preventing the first
despgraph eval to sync the ORCO layers. This is done to make
playback in solid mode faster.

When switching to material or render mode, the ORCO layers are not
masked anymore but there is no update tagging made in order to
generate them. This means the ORCO are only requested when
an update is manually made on the object (playback or edit).

This patch forces tagging of objects with deform modifiers when
switching to material or render shading mode. This is rather
crude and conservative heuristic which might impair performance
of switching between the modes. But at least, there is no missing
update.