UI: Improve the text feedback when translating
Previous text used abbreviations for text such as "D" which could mean distance or delta. transforms also gave the magnitude of the transform.

To improve understandability these abbreviations were expanded to give more meaning.
This also better matches the text for other operators such as scale.

The distance was also removed from single axis transformations as this is the same as the translation amount.