Transform: Remove unused "Bake Time" mode
Digging through the code, I couldn't find where this transform mode is
used.

From what I could see, it's supposed to be used in particle editing to
edit the `float *PTCacheEditKey::time` value of a particle.

In order to test the mode, I had to add a key map item to the
`3D View > Particle > Particle (Global)` keymap with the following
configuration:
- Event type: Keyboard `;` Press
- Identifier: transform.transform
  - Mode: Bake Time