Fix: Compositor redundantly executes if previews are enabled
The compositor executes redundantly if previews are enabled but no
other output is computed. This shouldn't be the case because previews
are secondary outputs that need a primary output to compute with.

We fix this by only executing the compositor if a primary output is
needed, regardless if previews are needed or not.