Add push constants uniform buffer fallback to the global descriptor set.
When descriptor_indexing is available, and a shader requires a push
constants fallback uniform buffer, we push a single push constant
that containing the binding location of the fallback buffer in the
global descriptor set. Shader code generation is changed to use this
push constant in order to resolve the PushConstants buffer accordingly.