Fix #140969: Joining armature objects that share their data crashes Blender
Add a precondition check that verifies all to-be-joined armatures are
unique. If this is detected, an error is shown and the operator is
aborted.

Joining shared Armatures is non-trivial, because the corresponding pose
bones may actually have different drivers or constraints on them
(because those live on the Object ID, and not the Armature). Duplicating
any shared Armature ID could be a first step, but I'm not convinced this
is always the right approach. This is why I just fixed the crash for
now. Users can then choose how to make each Armature unique (either by
deselecting objects or by duplicating their data).