Read asset index from hard-coded path in C++
Read the asset index file from a known file path. This removes the
assumption that the URLs in `_asset-library-meta.json` and
`_v1/asset-index.json` can double as file paths.

The `RemoteAssetListingDownloader` already processes
`_v1/asset-index.json` to replace the URLs with relative file paths, and
writes the result to into `_v1/asset-index.processed.json`. This is now
read by the C++ code as well.

This also has some refactors (naming variables `…url` instead of `…path`
where they contain a URL. And some changes in the (still too simplistic)
error reporting so that the actual filename is printed, of the file that
couldn't be read/processed.