WIP UI: SVG Icons as Default Mouse Cursors
This is work toward replacing our current fallback mouse cursors
(currently defined in wm_cursors.cc using bitmaps and masks that we
edit with a python program) with SVG sources that are rasterized when
needed. Like the current ones this only supplies aliased (1-bit)
cursors so just creates these with a brightness threshold. Each cursor
is just an SVG file just like the other icons but when defined also
uses hotspot and size values. The per-cursor size value needed so that
we can adjust their relative sizes regardless of the size in the SVG
files.