Fix #72913: Poly Build: "Face at Cursor Move" uses Ctrl which inverts snapping, disabling it undesirably
Poly Build is a tool that works well with snapping.

Therefore, hotkeys that disable snapping should be avoided in it.

This is not the case with the "Face at Cursor Move" operation of this
tool, which, since it is enabled with Ctrl, disables snapping.

To solve this problem, the snap property is now set to false in the key
map item of this operation. false because it is inverted to true.

The transform operator was edited to allow default snap values when a
property is not set.