Fix #140927: OpenGL: Annotate draw on surface broken
Recently we switched from Depth24S8 to Depth32FS8. On OpenGL the read
back of depth was not working as it asked the driver to download both
the depth and stencil part of the texture. This doesn't seem to be
supported and would not return any data.

This PR forces OpenGL to read back depth component when a depth/stencil
texture is used.

Might also fix other issues:
- [ ] #140912