Fix: Dead-lock in VSE audio waveform preview
The VSE preview job was pushing tasks while holding a lock
which is also used by the task. This could lead to a dead-lock
situation if scheduling decides to run the task immediately.

The easiest way to reproduce the issue is to run Blender with
`-t 1` command line argument and add WAV file.

The solution is to iterate over queue and create tasks while
holding a lock, but push them to the pool after releasing the
lock.

Pull Request: https://projects.blender.org/blender/blender/pulls/140972