Cleanup: Vulkan: Enable extensions
When the list of extensions is constructed for the vkCreateDevice it
used an old function that retrieved all extensions just to iterate to
check a specific extension was supported. However there is already a
list cached that is that is the subset of the desired extensions that
are supported by the device.

This cleanup will use that list instead of requering all supported
extensions.