Fix: Modifier Test: attributes change in MergedScrewWeld test object.
When re-saving the modifiers.blend test file the `UVMap` attribute of the
MergedScrewWeld test object is removed automatically, because the mesh
does not have any faces. This breaks the test because this attribute is
still expected on the modified object, which is generated by extruding the
original edge mesh.

Re-saving the file becomes necessary for adding more tests to it, but should
not break any of the existing tests. Update the test data here to ensure
future tests can be added without unrelated breakage.