Fix: Compositor: Potential crash if no active context is found
This is related to #140381, where the symptom of the bug was a crash
caused by an undefined behavior. In that case, setting a valid active
viewer key was the proper fix. However,
`find_active_context_recursive()` could return `nullptr` in theory so
the same problem might occur in the future.

The commit resolves the undefined behavior by avoiding the
dereferencing of a null pointer.