VSE: Add "Clear Keyframes" Operator to VSE Preview
This PR makes it easy to clear all keyframes of selected
strips by accessing just the Animation Menu in the Strip
menu and then clicking "Clear Keyframes...".

The feature is implemented by adding a keyframe_clear_vse
operator for the VSE as a counterpart to the keyframe_clear_v3d operator.