Fix #141190: keyed strip did not move to keyed position
The cause of the bug was that no cache invalidation of the strip took place
when a keyframe got inserted. The bug appeared because, in the commit that
created the bug, the line `scene->ed->cache_flag |= SEQ_CACHE_PREFETCH_ENABLE;`
was added, which automatically creates a cache for the strips that did not
get invalidated when a keyframe got inserted.

This solution adds logic to invalidate the strip's cache when a keyframe gets inserted.