Fix #141210: Do not check axis alignment for restoring auto perspective
For auto perspective option, it should always go back to perspective
when you exits from an axis aligned view or camera, previously it does
not handle axis aligned cases where there could be 45 deg rotation
present, in which case the viewport will not resotre to perspective when
you try to rotate afterwards. This fix removed the axis alignment check,
and always just ensures perspective if auto perspective is enabled,
because other view rotation operators will ensure snapping, so doesn't
seem to affect this.

Need more testing to ensure that this is working fine.