Fix #141012: Curve template allows zooming out beyond clip range.
The clipping flag of curve templates is applied only to point positions,
but allows zooming beyond the clip range due to numerial error.

If clipping is disabled the curve view rect is still restricted to
clip ranges (this could be considered a bug). However, zooming in and
back out leaves a tiny offset to the clip rect that allows one more
zoom-out step.

This patch ensures the curve rect is always within the clip rect.
When the difference is very small the view rect is snapped to the clip
rect.