Nodes: Allow auto linking of Extend sockets
The Make Links operator does not work with Extend Sockets and generally
do not respect the node's link function. This patch makes it such that
the operator will fallback to an Extend socket if all else fails,
connecting the first available unlinked output or fallback to the first
available output.

The main motivation for supporting this is to allow rapid connections to
the File Output node from big nodes like the Render Layers and images
nodes, which is typically a slow process for users.