Refactor: Replace 'ListBase' Mains by a VectorSet.
This is an unfortunate consequence of making Main a non-trivial C++
struct: it is no more guaranteed that next/prev pointers required for
ListBase remain immediately at the top of the struct anymore - and on
Windows they indeed don't.

So move to a more mordern form of storage for split-by-libraries mains.

WARNING: This is quite a sentivie change, extensively tested locally
(including library-related changes accross undos and redos, in complex
production scenes).