WIP: Proof of concept of get_transform and set_transform for RNA properties.
Proposed partial implementation for #141042, only handling Boolean properties
for now.

NOTE: Chose to go for also implementing this concept in C++ RNA code
itself, as it makes things easier, and imho ultimately is more
consistent. The new callbacks can only be defined for runtime properties
currently in C++ code, but nothing prevents it to be also possible for
'static' RNA ones if needs arise.