Fix: ctest prints command line that is not copy-pasteable
If you use BLENDER_VERBOSE=1, ctest prints out the blender command line
that was invoked, but it would fail if you just copy/pasted it into a
command prompt, because the "python-expr" argument contained spaces.
This fix surrounds the "python-expr" argument with quote marks, so you
can copy/paste it without modification.

Tested on Windows Command Prompt, Powershell and nushell.