Refactor and implement Rec.2100 PQ
The code is much more localized in the movie_write.cc.

The downside of this is that there are a bit of redundant color space
conversion happening. It seems that the better way for that would be
to somehow make BKE_imtype_requires_linear_float aware of FFmpeg HDR
output rather than adding a new argument. But if the latter is not
possible it is easy enough to restore the function argument from the
original patch from Aras.

The HLG code is based on OCIO configuration now. The code from Aras is
commented out. It is unclear how we can have a ground-truth to compare
the result with.