Fix #141253: EEVEE: Crash rendering a scene multiple times
This is an attempt to fix a race condition inside the
`Image` gpu texture cache. The `GPUTexture` class has a
refcounting mechanism used to avoid the `Image` datablock
freeing them mid rendering or right before rendering.

However, this mechanism is fragile and relied on the
assumption of no concurrent gpu texture request. This
was still the case in 4.4 because of the DRW global
lock. But this lock have been reduced in scope in 4.5.
This mean that texture creation and deletion are no
longer threadsafe.

The atomic refcount fixes the race condition in
`GPU_texture_free`.

Making the refcount atomic isn't enough and doesn't fix
the issue as the decrement (and free) can happen between
creation and increment. Instead we need to lock during the
creation + increment refcount (acquire) and the release
of the GPUTexture.