Refactor: Make file-list cache struct non-trivial C++ one
Makes it possible to use non-POD C++ types in `FileListEntryCache`. I'm
planning to use this for blender/blender!130543, to store status
information about downloaded previews. While it's possible to use POD
types instead, this is easier and a worthwhile general improvement.

I turned `FileList.filelist_cache` into an allocated pointer, so
`FileList` doesn't have to be ported to a non-POD type as well. We
probably want to do that too at some point.